DiscordBots/QC/requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
caer==1.9.8
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
discord.py==1.7.3
docutils==0.17.1
future==0.18.2
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==3.3
imageio==2.14.0
imageio-ffmpeg==0.4.5
jws==0.1.3
lxml==4.7.1
multidict==5.2.0
mutagen==1.45.1
mypy==0.812
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.2
oauth2client==3.0.0
opencv-contrib-python==4.5.1.48
pefile==2019.4.18
periodictable==1.5.3
Pillow==9.0.0
proglog==0.1.9
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.14.1
Pygments==2.9.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
pyperclip==1.8.2
python-decouple==3.5
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.3.1
termcolor==1.1.0
tqdm==4.62.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
websockets==10.1
yarl==1.7.2
yt-dlp==2022.2.4
pynacl